# docs/swagger.yaml
basePath: /
definitions:
...
  http.harborWebhook:
    properties:
      event_data:
        properties:
          repository:
            properties:
              date_created:
                type: integer
              name:
                type: string
              namespace:
                type: string
              repo_full_name:
                type: string
              repo_type:
                type: string
            type: object
          resources:
            items:
              properties:
                digest:
                  type: string
                resource_url:
                  type: string
                tag:
                  type: string
              type: object
            type: array
        type: object
      occur_at:
        type: integer
      operator:
        type: string
      type:
        type: string
    type: object
...
host: localhost:9300
info:
  contact:
    email: support@keel.sh
    name: Keel Support
    url: https://keel.sh
  description: Automated Kubernetes deployment updates. Keel is a tool for automating
    Kubernetes deployment updates.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://keel.sh/terms/
  title: Keel API
  version: "1.0"
paths:
...
  /v1/webhooks/github:
    post:
      consumes:
      - application/json
      description: Receives and processes GitHub webhook notifications for container
        registry events (both GitHub Packages and GitHub Container Registry)
      parameters:
      - description: GitHub event type (package or registry_package)
        in: header
        name: X-GitHub-Event
        required: true
        type: string
      - description: GitHub webhook payload
        in: body
        name: payload
        required: true
        schema:
          type: object
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
      summary: Trigger GitHub webhook
      tags:
      - webhooks
  /v1/webhooks/harbor:
    post:
      consumes:
      - application/json
      description: Receives and processes Harbor registry webhook notifications for
        image push events
      parameters:
      - description: Harbor webhook payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.harborWebhook'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
      summary: Trigger Harbor webhook
      tags:
      - webhooks
...
swagger: "2.0"